+++ app.env-example
@@ -1,6 +1,6 @@
 # environment variables for docker (app specific)
 DB_BACKEND=postgres
-POSTGRES_HOST=db
+POSTGRES_HOST=nlpdb
 NLP4ALL_ENV=production
 NLP4ALL_ADMIN_EMAIL=example@example.org
 NLP4ALL_ADMIN_PASSWORD=example
