+++ compose.development.yaml
@@ -10,6 +10,15 @@ services:
       - back-tier
     env_file:
       - ./db.dev.env
+  rabbitmq:
+    restart: always
+    image: rabbitmq:3.11
+    expose:
+      - "5672"
+    networks:
+      - back-tier
+    volumes:
+      - "rabbit:/var/lib/rabbitmq"
   app-dev:
     restart: always
     build: 
@@ -30,8 +39,28 @@ services:
       - back-tier
     depends_on:
       - db-dev
+      - rabbitmq
+      - worker-dev
+  worker-dev:
+    restart: always
+    build: 
+      context: .
+      target: nlpapp-dev
+      dockerfile: ./docker/python/Dockerfile
+    command: celery -A make_celery worker -l info
+    volumes:
+      - ./nlp4all:/home/app/nlp4all
+    env_file:
+      - ./db.dev.env
+      - ./app.dev.env
+    networks:
+      - back-tier
+    depends_on:
+      - db-dev
+      - rabbitmq
 networks:
   front-tier: {}
   back-tier: {}
 volumes:
   postgres_data_dev:
+  rabbit:
