+++ docker/nginx/nginx.conf
@@ -26,7 +26,8 @@ http {
             proxy_redirect off;
         }
 
-        location /static/ {
+        location /static {
+            include /etc/nginx/mime.types;
             alias /home/app/nlp4all/static/;
         }
