+++ make_celery.py
@@ -1,0 +1,4 @@
+from nlp4all import create_app
+
+flask_app = create_app()
+celery_app = flask_app.extensions["celery"]
