+++ nlp4all/controllers/data_source.py
@@ -1,9 +1,13 @@
 """Data sources."""  # pylint: disable=invalid-name
 
-# @todo: SKELETON, NOT IMPLEMENTED
-from flask import redirect, url_for
-
+import os
+from datetime import datetime
+from flask import redirect, url_for, current_app
+from werkzeug.utils import secure_filename
 from .base import BaseController
+from ..forms.data_source import AddDataSourceForm, ConfigureDataSourceForm
+from ..models import DataSourceModel
+from .. import db
 
 
 class DataSourceController(BaseController):  # pylint: disable=too-few-public-methods
@@ -14,19 +18,43 @@ class DataSourceController(BaseController):  # pylint: disable=too-few-public-me
     @classmethod
     def home(cls):
         """List of prepared data source"""
-        return cls.render_template("data_source_list.html", title="My Data Sources")
+        datasources = db.session.query(DataSourceModel).all()
+        return cls.render_template("data_source_list.html", title="My Data Sources", datasources=datasources)
 
     @classmethod
     def create(cls):
         """Upload / create page"""
-        return cls.render_template("create_data_source.html", title="New Data Source")
+        form = AddDataSourceForm()
+        if form.validate_on_submit():
+            f = form.data_source.data
+            filename = secure_filename(f.filename)
+            destination = os.path.join(current_app.root_path, 'data', filename)
+            # if the file exists we can add a timestamp to the filename
+            while os.path.exists(destination):
+                ts = datetime.now().strftime("%Y%m%d%H%M%S")
+                filename = f"{os.path.splitext(filename)[0]}_{ts}{os.path.splitext(filename)[1]}"
+                destination = os.path.join(current_app.root_path, 'data', filename)
+
+            f.save(destination)
+            ds = DataSourceModel(
+                data_source_name=form.data_source_name.data,
+                filename=filename)
+            db.session.add(ds)
+            db.session.commit()
+            return redirect(url_for("datasource_controller.configure") + f"/{ds.id}")
+        return cls.render_template("data_source_add.html", title="New Data Source", form=form)
 
     @classmethod
-    def configure(cls):
+    def configure(cls, datasource_id: int):
         """Specify data fields"""
+        ds = db.session.query(DataSourceModel).filter_by(id=datasource_id).first()
+        form = ConfigureDataSourceForm()
+        form.data_source_fields.choices = [(f, f) for f in ds.fields]
         return cls.render_template(
-            "configure_data_source.html",
-            title="Configure Data Source"
+            "data_source_configure.html",
+            title="Configure Data Source",
+            ds=ds,
+            form=form
         )
 
     @classmethod
