+++ nlp4all/routes/admin.py
@@ -19,3 +19,5 @@ AdminRouter.route("/manage_categories", methods=["GET", "POST"])(
     AdminController.manage_categories
 )
 AdminRouter.route("/add_org", methods=["GET", "POST"])(AdminController.add_org)
+AdminRouter.route("/celery_test/<int:x>/<int:y>", methods=["GET"])(AdminController.celery_test)
+AdminRouter.route("/celery_result/<string:task_id>", methods=["GET"])(AdminController.celery_result)
