+++ nlp4all/views/datasource/data_source_list.html
@@ -3,8 +3,8 @@
 <h1>My Projects</h1>
 <div class="content-section col-md-8">
 {% for d in datasources %}
-<a class="nav-item nav-link" href="#">{{ d.name }}</a>
-<p>{{ d.description }} </p>
+<a class="nav-item nav-link" href="#">{{ d.data_source_name }}</a>
+<p>{{ d.data_source_description }} </p>
 {% endfor %}
 </div>
 {% endblock content %}
