+++ requirements.txt
@@ -1,16 +1,23 @@
 alembic==1.9.1
+amqp==5.1.1
 asgiref==3.5.2
 astroid==2.14.2
+async-timeout==4.0.2
 attrs==22.1.0
 Babel==2.11.0
 bcrypt==4.0.1
+billiard==3.6.4.0
 blinker==1.5
 blis==0.7.9
 catalogue==2.0.8
+celery==5.2.7
 certifi==2022.12.7
 cffi==1.15.1
 cfgv==3.3.1
 charset-normalizer==2.1.1
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.2.0
 click==8.1.3
 colorclass==2.2.2
 confection==0.0.3
@@ -42,6 +49,7 @@ iniconfig==1.1.1
 itsdangerous==2.0.1
 Jinja2==3.1.2
 jsonschema==4.17.3
+kombu==5.2.4
 langcodes==3.3.0
 lazy-object-proxy==1.9.0
 Mako==1.2.4
@@ -56,13 +64,16 @@ oauthlib==3.2.2
 pathspec==0.10.3
 pathy==0.10.0
 pickleshare
+pika==0.13.0
 Pillow==9.3.0
 pip-upgrader==1.4.15
 platformdirs==3.0.0
 pluggy==1.0.0
 preshed==3.0.8
-psycopg[binary]==3.1.8
+prompt-toolkit==3.0.37
 psycopg-pool==3.1.6
+psycopg[binary]==3.1.8
+pyamqp==0.1.0.7
 pycparser==2.21
 pydantic==1.10.2
 PyJWT==2.6.0
@@ -75,11 +86,11 @@ requests-oauthlib==1.3.1
 requests==2.28.1
 six==1.16.0
 smart-open==5.2.1
-spacy==3.5.0
 spacy-legacy==3.0.12
 spacy-loggers==1.0.4
-SQLAlchemy==2.0.4
+spacy==3.5.0
 sqlalchemy-json==0.5.0
+SQLAlchemy==2.0.4
 sqlparse==0.4.3
 srsly==2.4.5
 terminaltables==3.1.10
@@ -93,8 +104,10 @@ typer==0.7.0
 typing_extensions==4.4.0
 tzdata==2022.7
 urllib3==1.26.13
+vine==5.0.0
 virtualenv==20.19.0
 wasabi==1.1.1
+wcwidth==0.2.6
 Werkzeug==2.2.3
 wincertstore==0.2
 wrapt==1.14.1
